+++ test/index.test.js
@@ -16,6 +16,7 @@ import { Request } from '@adobe/fetch';
 import { expect, use } from 'chai';
 import sinonChai from 'sinon-chai';
 import sinon from 'sinon';
+import AuthInfo from '@adobe/spacecat-shared-http-utils/src/auth/auth-info.js';
 import AccessControlUtil from '../src/support/access-control-util.js';
 
 import { main } from './utils.js';
@@ -76,6 +77,12 @@ describe('Index Tests', () => {
       },
       pathInfo: {
         suffix: '',
+      },
+      attributes: {
+        authInfo: new AuthInfo()
+          .withType('api-key')
+          .withAuthenticated(true)
+          .withProfile({ user_id: 'test-user' }),
       },
       env: {
         USER_API_KEY: apiKey,